+++ plugin/entra_company_agency/changelog.php
@@ -3,9 +3,12 @@
 <style type="text/css">@media screen and (max-width: 980px) {body{font-size: 18px}}</style>
 #################################<br />
 # @Name :  Plugin Release Notes   <br />
-# @Date : 03/07/2026             <br />
-# @Version : 1.0    	 	     <br />
+# @Date : 06/07/2026             <br />
+# @Version : 1.1    	 	     <br />
 #################################<br />
+<br />
+<u>Version 1.1 :</u><br />
+- Suit les changements de société : un changement de maison retire l'agence précédemment liée par le plugin et ajoute la nouvelle, sans jamais toucher une agence assignée manuellement<br />
 <br />
 <u>Version 1.0 :</u><br />
 - Ajoute une correspondance société Entra ID (companyName) / agence GestSup, appliquée automatiquement à chaque connexion SSO Office 365<br />